Refactor: Update admin panel components and dependencies
This commit updates Radix UI dependencies, renames several components for clarity and consistency, and removes unused sidebar store and team switcher components. These changes contribute to the ongoing refactoring of the admin panel.
