--- src/api/stores/sidebarStore.tsx
@@ -1,11 +1,0 @@
-import { create } from 'zustand';
-
-interface SidebarStore {
-  isOpen: boolean;
-  toggle: () => void;
-}
-
-export const useSidebarStore = create<SidebarStore>((set) => ({
-  isOpen: false,
-  toggle: () => set((state) => ({ isOpen: !state.isOpen })),
-}));
